Fix #4833 by suppressing trailers
In #4833, Moment was blocked from launching on
Linux because their `libsoup` clients hang when
talking to Envoy over HTTP/2 if the response
contains trailers, which is the case for JSON/gRPC
transcoded responses.

This commit adds a Lua filter to Envoy that
removes the unnecessary trailers that were causing
the hang, leaving no trailers and thereby no
`libcurl` bug-triggering content. Fixes #4833.

TESTED: added a new unit test to
  `local_envoy_test.py` that demonstrates that no
  trailers are sent on a JSON/gRPC transcoded
  response.