Inspect page overhaul (#4841)
* Introduce `Find()` method in the sidecar

The `Find()` method will allow the inspect page to
retrieve a range of actor IDs for a given state
type. This way we won't have to keep every actor
ID in memory.

Note that this PR doesn't yet using the pagination
capabilities of this method, but a follow-on PR
will, and it mades sense to write all the tests
and code in one go.

* `SidecarStateManager`: inspect page calls sidecar

Instead of using the in-memory state of
`SidecarStateManager` (which will be removed
soon), call the sidecar to get the current list of
actor IDs when required.

This lays the groundwork for removing the
in-memory actor state of `SidecarStateManager`.

* SidecarStateManager: remove 'self._actors'

This resolves the first half of the memory usage
issues seen in the facilitator (#4825): the Python
process no longer needs to keep every existant
actor ID in memory.

* Omit clearly unnecessary actor-change notifications

This reduces the load on the sidecar, as
known-to-have-already-existed actor changes don't
result in a re-query of all actor IDs at the
sidecar.

* Remove actor recovery.

- Remove `actors` from the `RecoverResponse` proto.
- Remove `RecoverActors()` method from `sidecar.cc`.
- Clean up related test code

This addresses the second part of the memory usage
issue described in #4825: the sidecar no longer
needs to load and communicate every existant actor
ID during recovery.

* Inspect: new methods for frontends to call

Before this commit, the inspect dashboard would
call `rpc GetAllStates`, and it would return all
all state data for all state IDs for all state
types. That wasn't very scalable.

The future dashboard will select a specific state
type and state ID to show, so it will do a more
efficient three-step process:
1. Fetch all state types - we introduce
   `rpc GetStateTypes` for this.
2. Fetch all state IDs for a single selected state
   type - we introduce `rpc GetStateIds` for this.
3. Fetch the data for a single selected state - we
   introduce `rpc GetState` for this.

This commit merely adds the new methods and their
tests; it doesn't yet reimplement the dashboard.

* Efficient three-column layout for the inspect page

Before this commit, the inspect page would fetch
and show all states in the application. This
wasn't scalable.

This commit introduces a three-column layout for
the inspect page. In the left column the user
selects a single state type. We then load state
IDs for only that state type. When the user
selects a state ID, we load the state data for
only that specific state. This new approach is
much more scalable and efficient.

As part of this refactor we remove the
`GetAllStates` RPC and related code. That's safe,
because the inspect frontend is served by the
application itself, so we're confident that
customers are always using the latest inspect page
(or, worst case, that they just need to refresh).

* Reduce application memory limit to 4Gi.

Now that actor IDs are no longer stored in memory,
the memory usage of facilitators is ~400MiB
instead of ~7GiB. This means we can reduce the
memory limit for customer applications again.

We prefer to have a slightly lower memory limit
for customer applications than 10GiB, because...
(1) we want to be able to schedule applications
    more efficiently.
(2) memory usage over ~1GiB has so far always
    indicated a bug that needed to be fixed.

* `sidecar.py`: apply suggestions from code review

Co-authored-by: Benjamin Hindman <benjamin.hindman@gmail.com>

* Backend fixes based on review feedback

* Refactoring of frontend based on review feedback

* Split `index.tsx` into multiple files

This is purely code moves, there's no logic
changes.

* Inspect: Sort "State Type" and "State ID" columns

Previously, the sort order of these columns
depended on the order in the response, which is
sorted only per consensus.

* Tasks inspector: link to states in states inspector

* Inspect: fix URL-was-getting-lost issue

Before this commit, refreshing an inspect page
with a url path like
`__/inspect/states/reboot.ping.Ping/example-pinger`
would reset that URL path to `__/inspect/states`.
This commit makes it so that the full URL path is
preserved on refresh.

* Fix type error

* Fix missed test.

---------

Co-authored-by: Benjamin Hindman <benjamin.hindman@gmail.com>