Various fixes to satisfy mypy (#4886)
These break the build in the Bazel 6 branch, for reasons that are
slightly obscure to me (possibly related is that we're pulling in an
updated types-protobuf package).

We need to make further changes to satisfy mypy in the Bazel 6 branch,
but some of those are closer to workarounds for mypy shortcomings.
Whereas these fixes are more clearly addressing errors, so seems useful
to land them before the rest of the Bazel 6 changes.