Upgrade to Bazel 6 (#4871)
* Upgrade to Bazel 6.5.0

This also requires upgrading to a newer version of protobuf
and gRPC, due to mutual dependencies between these packages.

This commit in isolation is broken, but will be fixed up later
in this PR.

* Move rules_buf patch to a patch

No need for a forked repo.

* Make mypy work with Bazel 6 (#4374)

* Move protoc-gen-mypy to public

* Apply patches to grpc and bazel-mypy-integration

* Apply mypy fixes

* Cleanup

* Fix tests

* Review

* Fix MacOS build and tests (#4379)

* Fix MacOS build and tests

* Double gRPC .so issue

* Add gRPC patch to fix macOS build

* Various fixes to satisfy mypy

Update / correct various type hints, and make control flow more
explicit in a few places where mypy wasn't able to infer the
correct properties in the past. It is slightly mysterious why
some of these places started breaking, but presumably the root
cause is updating mypy-protobuf, which brought in a new version
of the types-protobuf type hints.

Along the way, remove redundant `str` calls when using UUID
values in f-strings.

* Update goldens

* Update Pulumi

Previous versions of Pulumi had a dependency on an old version of
Protobuf; Pulumi >= 3.190 requires protobuf 5.x

Now that we have upgraded the version of protobuf we use (earlier
in this PR), upgrading the version of Pulumi is obligatory. In
principle, it would be possible to separate matters so that
Pulumi dependencies don't interact with Reboot's core dependencies,
but because we use Pulumi in some tests, that would not be possible
without a considerable refactor.

---------

Co-authored-by: Zakhar <zpilipuk@gmail.com>