Move buf dependencies to extra_deps()
`buf_dependencies` can access stuff from repos including the protobuf repo, so
to be safe we need to separate `buf_dependencies` into phase 2 of the dependency
fetching process. The current arrangement seems to not be broken with Bazel 6
for some reason, but it breaks on the Bazel 7 branch and is wrong in any event.

Along the way, move code that adds a repo to `repos.bzl`, instead of `deps.bzl`,
and add a missing call to `protobuf_deps()`.