Migrate from rules_proto to the built-in protobuf Bazel rules (#4942)
The former is deprecated. Unfortunately, not all of the functionality from
rules_proto has been moved to the protobuf rules yet; in particular,
`proto_descriptor_set` is not yet available. So we vendor a private copy of this
rule for the time being; once `proto_descriptor_set` has been added to upstream
protobuf we can get rid of this file:

https://github.com/protocolbuffers/protobuf/pull/23369

Note that we can't yet eliminate the dependency on rules_proto itself, because
other transitive dependencies of ours still use it. But we should be able to
drop it once we move to Bazel modules.