Send all traffic via local Envoy
This is a step towards horizontal scalability. By
routing all traffic via local Envoy, we can have
our routing logic in a single place, and we only
need to update one router (Envoy) to understand
that there can be multiple replicas.

The exception to all-traffic-via-Envoy is unit
tests in which there is only a single server
- which, to keep customer unit tests quick and
external-dependency-free, we make the new default.
In those cases, no Envoy is brought up and traffic
goes directly to the only server. All multi-server
tests (of which there are still plenty, and this
commit adds some more for additional coverage) now
use a local Envoy.