Fix routing bug in transaction `Watch()`
Before this commit, there was a bug in how
transaction participants would call `Watch()`: the
call did not carryr routing metadata (i.e. the
state ref header), which means that it relied on
the to-be-removed `DirectResolver` having
delivered a channel directly to the right server,
as opposed to e.g. to an Envoy that does routing.
If the channel was delivered to an Envoy, it would
have routed the call to a random server, which
would have assumed the call was for a previously
aborted transaction.

This commit fixes the bug in the following ways:
1. It adds the missing routing metadata to the
   `Watch()` call in transaction participants.
2. It adds a safety check in the server-side
   handling of `Watch()` calls to ensure that
   the call carries routing metadata; if not,
   the server rejects the call with an error.
3. It adds a safety check in the server-side
   handling of `Watch()` that the receiving server
   considers itself to be the right server for the
   coordinator state ref it was routed to (i.e. a
   sanity check that the routing happened
   correctly).