Introduce 'caller_id' for app-internal calls
Before this commit, applications would determine
whether a call was app-internal by inspecting the
`x-reboot-app-internal` header for a secret that
only the application itself would know. This had
three issues:

1. The secret had no way to be configured when
   there are multiple replicas of the same
   application.
2. There's a risk of a bugs that leaks the secret
   when the application thinks it'll call itself
   but actually calls another application.
3. It doesn't help with app-to-app traffic;
   applications that want to identify which
   application is calling them have no way to do
   so.

Other applications (notably our facilitator) would
look at the `x-forwarded-client-cert` header set
by Istio, which contains mTLS-checked information
about the caller. The problem with this approach
is that, when we have multiple replicas, we lose
this information when one of our replicas forwards
traffic to another replica in the same app.

This commit introduces the concept of 'caller_id',
in the `x-reboot-caller-id` header, which
identifies the calling application. This allows
applications to identify app-internal calls (i.e.
they see their own application ID as the caller
ID), as well as identify which other application
is calling them.

Subsequent commits focus on securing the
`x-reboot-caller-id` header in production
environments. This commit makes it work in dev,
replacing the old app-internal secret.