Fixes from #5028 (#5045)
* Health checks validate websocket server liveness

This commit adds a custom implementation of gRPC's
health check service. The custom servicer
transitively health checks the same process's
websocket server (in `react.py`). This
implementation would have caught the availability
issue in #5028.

This doesn't yet ensure that the health check is
actually performed; that's a subsequent commit.

* Cloud: perform liveness checks

As seen in #5028, our lack of liveness checks
before this commit was a problem in Reboot Cloud,
and in particular we need a liveness check that
verifies that websockets are working.

This commit relies on the custom implementation of
the gRPC health check servicer added in the
previous commit; it performs an additional
self-check on its websocket server before
reporting `SERVING`. We merely need to add
Kubernetes configuration to perform liveness
checks using the gRPC health check protocol.

* Update to latest `websockets`

In #5028 we suspected a bug in the `websockets`
library was causing availability issues for our
customer applications. While we don't know what
that bug might have been, updating to the latest
`websockets` version gives us the best chance of
avoiding recurrence.

The update is a major version bump, with some
small backwards incompatible API changes.

* Increase Groundcover trace retention to 7 days

In the course of debugging #5028 we learnt that
the default 24-hour retention policy for
Groundcover traces is insufficient for our needs.
This commit bumps the retention period to 7 days,
matching logs and metrics retention.

* Fix missed constant change, and unneeded import