Make workflows be static in Typescript generated code.
Typescript doesn't support 'abstract static' methods
yet, so we can't have a good type checking for static
workflow methods. As a workaround we check that all
workflows are static in the servicer constructor,
since we know that we instantiate servicers early.
To not break Moment.dev with this change we still
allow non-static workflows in servicers, but log
a warning that they need to migrate to static workflows.