Don't cache all idempotent mutations
This is a stepping stone to relying on the bloom filter and recovering
any idempotent mutations as necessary. In a subsequent commit we will
first recover just the bloom filter and only later specific idempotent
mutations as necessary.

We probably do want to reintroduce caching of idempotent mutations in
the future, as well as recovering idempotent mutations in batch so
that we better support replays, but for simplicity we'll do that
later.