Release from `manylinux` to increase `glibc` compatibility to `2.34` (from `2.35`) (#5131)
* Release from `manylinux`

This commit adds a new `manylinux-builder`
container that we use during the release to
publish our wheel file. That ensures that the
binaries we package into our wheel file are built
on a `glibc` that's compatible with e.g. Vercel's
builder machines (Amazon Linux 2023, `glibc`
version `2.34`).

To ensure that these released binaries will work,
we also test them on every PR: we add the
`manylinux-runner` container to run a single
representative test using the binaries
built-as-they-would-be-released.

* Reduce code duplication in our GitHub `.yml` files

This moves a few shared parts of `respect.yml` and
`manylinux.yml` to shared actions, reducing code
duplication. It doesn't change any functionality.

* Rename `respect.yml` to `devcontainer.yml`

The name `respect.yml` was a leftover of a bygone
era; the more accurate name in this day and age is
"devcontainer" to indicate it runs code in our
devcontainer. Compare this to `manylinux.yml`
which runs code in our `manylinux` container.

* Reorganize pip package building

In this PR, the `//reboot:reboot` target becomes
buildable only on our `manylinux` containers - not
on our workstations. This commit addresses several
problems that arose from that:
1. A `bazel build //reboot/...` will no longer
   include `//reboot:reboot` (by tagging it as
   `"manual"`), so that we can continue to use
   wildcards on our workstations. We build the
   `//reboot:reboot` target in the `manylinux`
   container in CI to ensure it works.
2. The `//reboot:reboot.dev` target is no longer
   given a different `*.dev0` version number - it
   now has the same version number as
   `//reboot:reboot`, but (as before) will NOT be
   run through `auditwheel`, meaning it's useful
   for local development but not release. Making
   the version numbers the same allows it to be
   used by e.g.
   `//reboot/containers/reboot-base:dev`.

To make this possible, we have to reorganize the
building of the pip package in `pip_package.bzl` a
bit: we make `//reboot:reboot` depend on the
output of `//reboot:reboot.dev`.

* Update formatting to match `clang-format` 20

With the move to `clang-format` version 20,
apparently the default styles we're building off
have had some changes. This commit makes all of
the formatting changes needed to make the new
`clang-format` happy (and they seem like
improvements!).

* Address review comments

* Further review feedback