Remove bearerToken from cache key hash
Previously, the bearerToken was included in the cache key hash,
which caused unnecessary caching because the bearerToken
is refreshed often AND each time the page loads.

This was done for security reasons, we didn't want to expose data that
was previously only available to logged in users to unauthenticated users.
However, since the cache is stored in plaintext on the client, anything
in the cache is always accessible to any user, so this feature wasn't
providing security and was leading to cache misses.

Also, create requestBearerTokenHash which includes the bearerToken
for tracking unused readers, as this functionality breaks without the
bearerToken.